s3:smbd: change user_struct->vuid to uint64_t
Only sconn->smb1.sessions.next_vuid remains as uint16_t,
so that we do not generate larger values yet.

metze

Autobuild-User: Stefan Metzmacher <metze@samba.org>
Autobuild-Date: Wed Jun  6 12:07:33 CEST 2012 on sn-devel-104

+++ source3/smbd/globals.h
@@ -476,7 +476,7 @@ struct pending_auth_data;
 
 struct user_struct {
 	struct user_struct *next, *prev;
-	uint16 vuid; /* Tag for this entry. */
+	uint64_t vuid; /* Tag for this entry. */
 
 	char *session_keystr; /* used by utmp and pam session code.
 				 TDB key string */
@@ -590,7 +590,7 @@ struct smbd_server_connection {
 			 *  Set by us for CORE protocol.
 			 */
 			int max_send;
-			uint16_t last_session_tag;
+			uint64_t last_session_tag;
 
 			/*
 			 * this holds info on user ids that are already
